Swap stuff over to a utilities namespace
