+++ Assets/Scripts/Utilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be425c0e9ed2719478a8ceef636537ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
