+++ Assets/Scripts/Utilities/StringUtilities.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a45a2a3f6d7f7c7489d76271eb7859e0
+guid: 7ee916e66cf4df6449cde65a7b2461c9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
